chore: move token utils
