+++ src/general_harvester.py
@@ -22,7 +22,7 @@ from src.utils import confirm_transaction
 from src.utils import get_abi
 from src.utils import get_hash_from_failed_tx_error
 from src.utils import get_last_harvest_times
-from src.utils import get_token_price
+from src.token_utils import get_token_price
 from src.utils import send_error_to_discord
 from src.utils import send_success_to_discord
